backend/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
requests==2.24.0
Flask-SQLAlchemy==2.5.1
Jinja2==3.0.3
marshmallow==3.14.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
SQLAlchemy==1.4.32
MarkupSafe==2.0.1
itsdangerous==2.0.1
marshmallow-sqlalchemy==0.27.0
uWSGI==2.0.20
